Barnsley Council Initial Clean and Vis
After reviewing and discussing the report with shareholders and journalists of the local community page, I was tasked with communicating the data effectively. This was the start. It includes line chards and also bar charts demonstrating spending per Expense Type. The initial visualisations are to confirm and communicate findings in a simple chart.